lib/streamaggr: reduce number of inuse objects
The main change is getting rid of interning of sample key.
It was discovered that for cases with many unique time series aggregated by vmagent
interned keys could grow up to hundreds of millions of objects.
This has negative impact on the following aspects:
1. It slows down garbage collection cycles, as GC has to scan all inuse objects periodically.
The higher is the number of inuse objects, the longer it takes/the more CPU it takes.
2. It slows down the hot path of samples aggregation where each key needs to be looked up
in the map first.

The change makes code more fragile, but suppose to provide performance optimization for
heavy-loaded vmagents with stream aggregation enabled.

Signed-off-by: hagen1778 <roman@victoriametrics.com>