Fix data inconsistency in dateMetricIDCache
Data inconsistency manifests when the cache is used concurrently. After an entry
is added to cache, it is possible that subsequent entry retrievals may result in
cache miss (i.e. the entry is not found in cache even though it is there).

If an entry is not found in immutable map, the caller acquires the lock to
search the mutable map. But it is possible that the entry is moved from the
mutable map to the immutable map before that lock is acquired.

Signed-off-by: Artem Fetishev <wwctrsrx@gmail.com>