Fix inconsistent error handling in Storage.AddRows()
Storage.AddRows() returns an error only in one case: when
Storage.updatePerDateData() fails to unmarshal a metricNameRaw. But the same
error is treated as a warning when it happens inside Storage.add() or returned
by prefillNextIndexDB().

This commit fixes this inconsistency by treating the error returned by
Storage.updatePerDateData() as a warning as well. As a result Storage.add() does
not need a return value anymore and so doesn't Storage.AddRows().

Additionally, this commit adds a unit test that checks all cases that result
in a row not being added to the storage.

Please note that this commit does not attempt to fix the correctness of the
TotalRowsAdded counter metric in Storage.AddRows(). After fixing the error
handling its value may be inaccurate but so was it before this fix. A separate
PR will be sent to fix this metric and also add more metrics to provide more
visibility into what went wrong during adding records to the storage.

Signed-off-by: Artem Fetishev <wwctrsrx@gmail.com>