Allow to disable fsync when running unit tests
fsync() ensures that the data is written to disk. In production this is needed
for data durability. However, during the development, when the unit tests are
run, this level of durability is not needed. Therefore fsync() can be disabled
which will decrease time waiting for unit tests to complete. By 25%, to be
exact.

The disabling is done by setting the DISABLE_FSYNC_FOR_TESTING environment
variable. The valid values for this variable are the same as the values of the
arg of `go doc strconv.ParseBool`:

```
1, t, T, TRUE, true, True, 0, f, F, FALSE, false, False.
```

Any other value means `false`.

The variable is set for the `make test` build target as this is the target that
is mandatory to run by a developer before pushing the code. If you want to run
other build targets with fsync disabled, the variable can be set like this:

```shell
DISABLE_FSYNC_FOR_TESTING=1 make test-full
```

or

```shell
DISABLE_FSYNC_FOR_TESTING=1 go test ./lib/storage
```

Disabling fsync() is intended for testing purposes only and the name of the
variables reflects that.

Signed-off-by: Artem Fetishev <wwctrsrx@gmail.com>