Also disable directory sync for testing
This results in the major descrease of test running time, x2 faster.

Since disabling file sync and directory sync is done in different packages
(lib/filestream and lib/fs) the code that reads the environment variable is
moved to a separate package for reuse.

Also the fsync has been disabled for all test build targets.

Signed-off-by: Artem Fetishev <wwctrsrx@gmail.com>