RegisterMetricNames does not create new per-day index entries for new dates for existing timeseries
Once the timeseries is in tsidCache, new entries won't be created in per-day index because the
RegisterMetricNames() code does consider different dates for the same
timeseries. So this case has been added.

The same bug exists for AddRows() but it is not manifested because the index
entries are finally created in updatePerDateData().

RegisterMetricNames also updated to increase the newTimeseriesCreated counter
because it actually creates new time series in index.

A unit test has been added that tests all possible data patterns (different
metric names and dates) and code branches in both RegisterMetricNames and
AddRows.

Signed-off-by: Artem Fetishev <wwctrsrx@gmail.com>