Do not drop samples with stale NaN value for time series that hasn't been yet
This patch reverts the fix for this issue:
https://github.com/VictoriaMetrics/VictoriaMetrics/issues/5069

After discussing it we've come to conclusion that this is a valid behavior which
can be avoided by deleting the time series only once the corresponding stale
NaNs have been received.

On the other hand, the fix leads to lost stale NaNs in some rare but valid use
cases. For example:

-   In a cluster configuration the samples for a given time series are normally
    sent to the same vmstorage replica. However, wminsert may reroute the
	samples to another replica because the original one is down or is
	overloaded. In this case the stale NaN may end up on a replica that has no
	data for that time series, but we still want to record that sample.

Thus, reverting that fix.

Signed-off-by: Artem Fetishev <wwctrsrx@gmail.com>