lib/promscrape/discovery/kubernetes: re-use http.Client for different group watchers
Creating a separate http.Client for every groupWatcher results into maintaining a separate connections pool per client. Thus this increases amount of opened connections and can lead to port exhaustion in cases when there are a lot of unique groupWatchers created.

As a first step to address that, this commit adds re-use of http.Client for groupWatchers when this is possible. This is supposed to lower amount of open connections an improve connection pooling efficiency.

See: https://github.com/VictoriaMetrics/VictoriaMetrics/issues/5971
Signed-off-by: Zakhar Bessarab <z.bessarab@victoriametrics.com>