lib/promscrape: Fix TestClientProxyReadOk flaky test
For hijacked connections, one has to read from the connection buffer, but still
write directly to the connection. Otherwise, when reading directly from such
connections, the first byte may be lost. This, in turn corrupts the ClientHello
TLS handshake message and when the backend server receives it, it closes the
connection and reports the following error in the log:

```
http: TLS handshake error from 127.0.0.1:33150: tls: first record does not look
like a TLS handshake
```

The first byte may be lost because underlying HTTP request handler may read it
from the connection and put it into the buffer. As the result, subsequent
connection reads won't see that byte.

-   See: https://github.com/golang/go/issues/27408
-   The fix is taken from : https://github.com/k3s-io/k3s/pull/6216

Signed-off-by: Artem Fetishev <rtm@victoriametrics.com>