lib/flagutil/dict: properly update default value in case there is no key value set
If a dict flag has only one value without a prefix it is supposed to replace default value.

Previously, when flag was set to `-flag=2` and the default value in NewDictInt was set to 1 the resulting value for any flag.Get() call would be 1 which is not expected.

This commit updates default value for the flag in case there is only one entry for flag and the entry is a number without a key.

Signed-off-by: Zakhar Bessarab <z.bessarab@victoriametrics.com>