apptest: typical cluster configuration for business logic tests
Add the ability to create a simple cluster configuration for tests that do not verify the cluster-specific behavior but instead are focused on the business logic tests, such as API surface or MetricsQL. For such tests this cluster configuration will be enough in most cases.

Cluster-specific tests should continue creating custom configurations.

Signed-off-by: Artem Fetishev <rtm@victoriametrics.com>