app/vmauth: allow to start with empty auth config file
 This commit adds ability to launch vmauth without configuration file.
Which is possible use case for operator based installations.

  Operator provides global resource `VMAuth` and allows to create
`VMUser` objects for it. Eventually operator creates configuration for
`VMAuth` based on user defined selectors for `VMUser`.

  Since there is no direct relations between
those objects. And any object could be created in on-demand by Kubernetes users.
It's required to be able to start `vmauth` with empty auth config file.

Related issue:
https://github.com/VictoriaMetrics/VictoriaMetrics/issues/6467