Remove legacy copies of add() and registerMetricNames()
The plan was to keep them in order to be able to switch between implementations
using a flag. But it may be very difficult to keep them both. The new approach
is to remove all legacy code that is not needed and test forward compatibility
using an integration test that first starts an old implementation and then new
impl on top of data produced by the old one.

Signed-off-by: Artem Fetishev <rtm@victoriametrics.com>