app/vmalert/notifier: fix rendering of Alertmanager notification body
c7fc0d0d2fa5c56bff5731709c3067532edda386 Commit enabled skipping alerts in case there is no labels present for an alert. This made clause which was adding a comma for the JSON list incorrect as it is not possible to determine if the next alert will be skipped or not.

This fix renders all alert labels in advance allowing properly format JSON payload for Alertmanager notification.

Signed-off-by: Zakhar Bessarab <z.bessarab@victoriametrics.com>