app/vmselect/promql: account for staleness when populating realPrevValue
When vmselect process a rollup function it fetches all the raw samples on
requested `start-end` interval of the query. It then loops through the raw samples,
picks the range of the samples based on provided `step` interval and invokes a rollup
function for each of the picked ranges of samples.

During this processing, vmselect always populates the `realPrevValue` field with the
closest previous raw sample value before the picked range of samples. This `realPrevValue`
is used by rollup functions like increase_pure or delta to decide whether the counter
change happened or not. For example, we get the counter value == 1. If we've seen this
counter before and its value was also 1 - then no change happened. If we didn't see it
before, then this counter should have started with value=0 and we need to account for
`1-0=1` change. All this is required to deal with situations when scrapes are missing
or `step` is too small.

However, vmselect doesn't check how "old" is the `realPrevValue`. In other words, it
doesn't respect the staleness interval when picking it. In result, depending on the
`start` and `end` params, vmselect can use `realPrevValue` which is a couple of hours
old and is unlikely to be a temporary scrape fail. In result, some increases can
be incorrectly ignored by vmselect.

This change makes sure that vmselect doesn't populate `realPrevValue` with samples
that are older than staleness interval.

Signed-off-by: hagen1778 <roman@victoriametrics.com>