metricsql: bump to v0.83.0
See https://github.com/VictoriaMetrics/VictoriaMetrics/issues/7703

The update also returns an error if metric name is specified twice in metrics selector.
For example, `foo{__name__="bar"}` is not allowed anymore. It would successfully parse before
this change, but it won't satisfy the search filter any way. So it had no sense in supporting this. This is why some test cases were removed.

Signed-off-by: hagen1778 <roman@victoriametrics.com>