export API: cancel export process on client connection close
At this time bufferedwriter silently ignores connection close errors. It
may be very convenient in some situations (to not log such unimportant
errors), but it's too implicit and unsafe for the others. For example,
if you close export API client connection in the middle of communication,
VictoriaMetrics won't notice it and will start to hog CPU by exporting
all metrics into nowhere until it process all of them. If you'll make a
few retries, it will be effectively a DoS on the server.

This commit replaces this implicit error suppressing with explicit error
handling which fixes the issue with export API.